functions/create-google-doc/requirements.txt
functions-framework==3.*
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-firestore
google-cloud-secret-manager